implement save
save() uses the libfdt sw (sequential write) interface to stream out the
FDT pieces. This is quite inefficient as it requires at least twice as
much memory as the devicetree nodes and properties, and makes a complete
copy of everything.

We could improve this situation by changing save to write to an output
iterator, but that would mean rewriting a bunch of libfdt code.