add missing non-member accessor functions
Previously I decided not to duplicate member accessors as non-member
functions.

This leads to code which needs to static_cast<T &> to access the insides
of a reference_wrapper<T>. Adding non-member overloads for accessors
nicely cleans this up as reference_wrapper<T> implicitly converts to T&.
It also has the added benefit of making the public interface to libfdt++
consistent.