fix CI trying to rebuild dtbs
When git clones a repository it sets the file time of newly checked out
files to the current system time.

Because we include pre-built dtbs in the git tree and don't want to
rebuild them during CI we need to touch the pre-built files before
running the tests. That way make won't try to rebuild them.